add hook `useAudioPlayerCurrentTime`
